requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.3
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
idna==3.6
packaging==24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.2.1
django-filter==24.2
inflection==0.5.1
google-auth==2.32.0